requirements.txt
pandas
requests
python-dotenv
psycopg2-binary
mediacloud
rich
newspaper3k
beautifulsoup4
transformers
matplotlib
seaborn
aws-cdk-lib
constructs
tqdm
torch
accelerate
